# openapi-agent-v2.yaml
openapi: 3.1.0
info:
  title: TuxBridge Agent Session API
  version: 0.3.0-alpha
  description: >-
    Session-first agent control surface. A GPT discovers a workspace, selects it to
    create a readable session reference, records task summary/plan/todos, and uses
    that session reference for all subsequent agent operations.
servers:
  - url: https://tuxbridge.example.com
    description: Replace with the HTTPS endpoint fronting TuxBridge.
security:
  - bearerAuth: []
...
paths:
...
  /v1/agent/commands/run:
    post:
      tags: [Agent tools]
      operationId: runCommand
      summary: Run a command in the session workspace.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RunCommandRequest' }
      responses:
        '200': { $ref: '#/components/responses/OperationResponse' }
...
  /v1/agent/git/sync:
    post:
      tags: [Agent tools]
      operationId: gitSync
      summary: Fetch, fast-forward pull, or push for the session workspace.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GitSyncRequest' }
      responses:
        '200': { $ref: '#/components/responses/OperationResponse' }
...
components:
  responses:
...
    OperationResponse:
      description: Durable frozen operation state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Operation' }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Operation:
      type: object
      required: [id, session, workspace, tool, args, policy, status, created_at_unix_ms, updated_at_unix_ms]
      properties:
        id: { type: string }
        session: { type: string }
        workspace: { type: string }
        tool: { type: string }
        args: {}
        policy: { type: string, enum: [automatic, confirm, blocked] }
        status: { type: string, enum: [pending_approval, executing, completed, failed, denied, blocked] }
        created_at_unix_ms: { type: integer }
        updated_at_unix_ms: { type: integer }
        result: {}
        error: { type: [string, null] }
...
    RunCommandRequest:
      type: object
      required: [session, argv]
      properties:
        session: { type: string }
        argv:
          type: array
          minItems: 1
          items: { type: string }
        timeout_seconds: { type: [integer, 'null'], minimum: 1 }
        background:
          type: boolean
          default: false
          description: Start as a background job when true.
...
    GitSyncRequest:
      oneOf:
        - type: object
          required: [action, session]
          properties:
            action: { const: fetch }
            session: { type: string }
        - type: object
          required: [action, session]
          properties:
            action: { const: pull }
            session: { type: string }
        - type: object
          required: [action, session, remote, branch]
          properties:
            action: { const: push }
            session: { type: string }
            remote: { type: string }
            branch: { type: string }
...